Synchronize in-app polls / prevent pile-ups
Prevents multiple poll requests from piling up in the event of a
slowdown and/or too-short polling interval. Only one poll dispatch
can be active at once; if another poll is attempted while one is
already active, it will just drop. Another dispatch will happen at
the next polling interval anyway.

This is really an edge case after the previous updates to address
ack and user callback ordering. I encountered it when pausing in
the debugger - one thread was paused while others continued. That
resulted in a bunch of poll requests queuing up in the network
executor, which all ran in rapid succession after resuming.

The same could potentially happen in the real world in the event
of network slowdown/delay, especially if the polling interval was
configured to be short. So this update prevents that whole issue.