Notif helpers can boot their own Snapyr instance
tl;dr fixes crash/non-displayed notifications when a Snapyr notif
arrives on device after app has closed, in certain situations
including use in React Native apps.

Scenario: a notification arrives, or is tapped, and the
notification helper code no longer has access to the Snapyr
instance.

This can happen if Snapyr is not instantiated early enough in
application lifecycle to be available by the time the notification
service or notification listener activity need to use it. (this is
the case for React Native, where we can't initialize Snapyr until
the main activity starts up)

To account for this, the main Snapyr init code now stashes minimal
configs necessary to be able to communicate with the back end -
namely, the write key and environment.

Then, in notification context, if there is not a valid Snapyr
instance, it will use a helper to instantiate a one-off, minimal
"helper" Snapyr instance. This disables things like lifecycle
listeners, and uses its own one-off memory queue, so that
notification tracking data can be sent to Snapyr without affecting
anything about the main Snapyr instance that may be booted up soon
afterward.