Allow notif icon to be configured by client app
This is done by convention rather than by passing an explicit
config to Snapyr - partly because other notification services do
it this way; partly so that notification code can find the icon
without needing to rely on communicating config with the main app.

If client app contains a drawable or mipmap named
`ic_snapyr_notification_default`, this will be used. If not, look
for a drawable or mipmap named `ic_notification` (a common name
for notification toolkits).

If neither is found, falls back to the app's launcher icon (which
is what we used for all notifications previously); if even that is
not found, finally, use a built-in Android icon.