Notification tap behavior improvements
TaskStackBuilder is generally recommended for setting up the
activity stack for a notification, but it makes it harder to
control the opening behavior with our tracking activity in the
middle.

The previous setup was causing a full new activity to be launched
even if the activity was in `singleTask` (or other "single...")
launch mode and already open.

Moves some of the responsibility for opening the "real" activity
to the tracking activity, which can set flags on the intent to
improve this behavior.

In this update, if there is no deep link URL, the same activity
that was already open should simply come to the foreground (note
that this does NOT trigger any lifecycle calls on the activity,
meaning there is not an obvious way for user code to know that it
was brought to the foreground from a notification tap at this
time).

If there is a deep link URL, a new version of the activity will be
launched with that data, which is the generally expected behavior.

For a `singleTask` activity, the task will be brought to the
foreground and the `onNewIntent()` method will be triggered with
the new data from the notification.