Fix for flush called immediately after track, etc.
Track/identify/etc. calls add events to the queue on executor, but
flush was calling sendQueue.flush directly/immediately. That could
result in sendQueue.flush being run before the event reached the
queue, causing the flush to miss that event.

This moves the sendQueue.flush call to the executor as well, so it
will run sequentially after any previously-called event queuing
methods.

Fixes push tracking events not making it to back end when a
notification was tapped after the application was closed