ServiceFacade activity reference improvements
Set current activity onActivityResume, and set regardless of
whether the rest of this method (and onActivityCreated) is to be
skipped. Ensures the current activity is set accurately when
switching between activities that are not newly-created, i.e.
when going "back" from one activity to another.

Use WeakReference to prevent memory leaks or improper interruption
to activity lifecycle.

Null check before attempting to open in-app overlay, as it will
crash if the activity can't be found.