Fix - text-only notifs expandable on some OSes
Some Android devices/OSes, including Google Pixel, automatically
make long text notifications expandable, but others do not. In that
case notifications with text that exceeds the width of the notification
will have the text truncated with an ellipsis, with no way to expand
the notification to see more.

Setting `BigTextStyle` ensures that notifications with longer text
can be expanded.

This applies only to notifications that do not include an image.
Notifs w/ image will override this with a `BigPictureStyle` which
includes the image, and a length-limited summary text. Note that
BigPictureStyle does NOT allow the text to be expanded, so notifs
with images can only display the ellipsis-truncated text alongside
the image.