Use plain WebViewClient (no override for now)
InAppWebviewClient `shouldOverrideUrlLoading` is currently unused,
as the overlay script cancels clicks on <a> tags and sends event
data to the SDK instead for handling.

Existing code in InAppWebviewClient could cause error/crash if the
overlay script doesn't work for some reason, so disabling it to be
safe. We may want to update this and use again in the future, so
leaving the class here